+++ src/ip6.c
@@ -3,7 +3,7 @@
  *
  * Copyright (c) 2002 Dug Song <dugsong@monkey.org>
  *
- * $Id: ip6.c,v 1.1 2002/11/27 06:42:20 dugsong Exp $
+ * $Id: ip6.c,v 1.2 2005/01/23 07:36:54 dugsong Exp $
  */
 
 #include "config.h"
@@ -25,7 +25,7 @@ ip6_checksum(void *buf, size_t len)
 	nxt = ip6->ip6_nxt;
 	
 	for (i = IP6_HDR_LEN; IP6_IS_EXT(nxt); i += (ext->ext_len + 1) << 3) {
-		if (i >= len) return;
+		if (i >= (int)len) return;
 		ext = (struct ip6_ext_hdr *)((u_char *)buf + i);
 		nxt = ext->ext_nxt;
 	}
